requirements.txt
autopep8==1.6.0
click==8.0.3
deco==0.6.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
toml==0.10.2
Werkzeug==2.0.2
